Remove old pep classes. Remove mixnet package. Add documentation
