+++ src/test/kotlin/org/cryptobiotic/pep/MixnetDecryptorTest.kt
@@ -1,4 +1,4 @@
-package org.cryptobiotic.mixnet
+package org.cryptobiotic.pep
 
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.unwrap
@@ -6,7 +6,6 @@ import electionguard.core.ElGamalCiphertext
 import electionguard.core.ElementModP
 import electionguard.core.productionGroup
 import electionguard.publish.makeConsumer
-import org.cryptobiotic.pep.CiphertextDecryptor
 import org.cryptobiotic.verificabitur.bytetree.ByteTreeNode
 import org.cryptobiotic.verificabitur.bytetree.readByteTreeFromFile
 import kotlin.test.Test
